Change MODIS by adding more options

+++ Collect/MCD43/DataAccess.py
@@ -22,11 +22,12 @@ import requests
 from joblib import Parallel, delayed
 
 # Water Accounting modules
+import wa
 import wa.General.raster_conversions as RC
 import wa.General.data_conversions as DC
 from wa import WebAccounts
 
-def DownloadData(Dir, Startdate, Enddate, latlim, lonlim, Waitbar, cores):
+def DownloadData(Dir, Startdate, Enddate, latlim, lonlim, Waitbar, cores, hdf_library, remove_hdf):
     """
     This function downloads MCD43 daily data
 
@@ -73,32 +74,11 @@ def DownloadData(Dir, Startdate, Enddate, latlim, lonlim, Waitbar, cores):
     if not os.path.exists(output_folder):
         os.makedirs(output_folder)
 
-    # Download list (txt file on the internet) which includes the lat and lon information of the MODIS tiles
-    nameDownloadtext = 'https://modis-land.gsfc.nasa.gov/pdf/sn_gring_10deg.txt'
-    file_nametext = os.path.join(output_folder, nameDownloadtext.split('/')[-1])
-    try:
-        try:
-            urllib.urlretrieve(nameDownloadtext, file_nametext)
-        except:
-            data = urllib2.urlopen(nameDownloadtext).read()
-            with open(file_nametext, "wb") as fp:
-                fp.write(data)
-    except:
-        from requests.packages.urllib3.exceptions import InsecureRequestWarning
-        requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
-        with open(file_nametext, "wb") as fp:
-            data = requests.get(nameDownloadtext, verify=False)
-            fp.write(data.content)
-
-    # Open text file with tiles which is downloaded before
-    tiletext=np.genfromtxt(file_nametext,skip_header=7,skip_footer=1,usecols=(0,1,2,3,4,5,6,7,8,9))
-    tiletext2=tiletext[tiletext[:,2]>=-900,:]
-
-    # This function converts the values in the text file into horizontal and vertical number of the tiles which must be downloaded to cover the extent defined by the user
-    TilesVertical, TilesHorizontal = Tiles_to_download(tiletext2=tiletext2,lonlim1=lonlim,latlim1=latlim)
+    # Define which MODIS tiles are required
+    TilesVertical, TilesHorizontal = wa.Collect.MOD15.DataAccess.Get_tiles_from_txt(output_folder, hdf_library, latlim, lonlim)
 
     # Pass variables to parallel function and run
-    args = [output_folder, TilesVertical, TilesHorizontal, lonlim, latlim]
+    args = [output_folder, TilesVertical, TilesHorizontal, lonlim, latlim, hdf_library]
     if not cores:
         for Date in Dates:
             RetrieveData(Date, args)
@@ -110,15 +90,17 @@ def DownloadData(Dir, Startdate, Enddate, latlim, lonlim, Waitbar, cores):
         results = Parallel(n_jobs=cores)(delayed(RetrieveData)(Date, args)
                                          for Date in Dates)
     # Remove all .hdf files
-    os.chdir(output_folder)
-    files = glob.glob("*.hdf")
-    for f in files:
-        os.remove(os.path.join(output_folder, f))
+    if remove_hdf == 1:
+        os.chdir(output_folder)
+        files = glob.glob("*.hdf")
+        for f in files:
+            os.remove(os.path.join(output_folder, f))
 
-    # Remove all .txt files
-    files = glob.glob("*.txt")
-    for f in files:
-        os.remove(os.path.join(output_folder, f))
+        # Remove all .txt files
+        files = glob.glob("*.txt")
+        for f in files:
+            os.remove(os.path.join(output_folder, f))
+
 
 	return results
 
@@ -132,11 +114,11 @@ def RetrieveData(Date, args):
     args -- A list of parameters defined in the DownloadData function.
     """
     # Argument
-    [output_folder, TilesVertical, TilesHorizontal, lonlim, latlim] = args
+    [output_folder, TilesVertical, TilesHorizontal, lonlim, latlim, hdf_library] = args
 
     # Collect the data from the MODIS webpage and returns the data and lat and long in meters of those tiles
     try:
-        Collect_data(TilesHorizontal, TilesVertical, Date, output_folder)
+        Collect_data(TilesHorizontal, TilesVertical, Date, output_folder, hdf_library)
     except:
         print "Was not able to download the file"
 
@@ -160,42 +142,7 @@ def RetrieveData(Date, args):
 
     return True
 
-def Tiles_to_download(tiletext2,lonlim1,latlim1):
-    '''
-    Defines the MODIS tiles that must be downloaded in order to cover the latitude and longitude limits
-
-    Keywords arguments:
-    tiletext2 -- 'C:/file/to/path/' to path of the txt file with all the MODIS tiles extents
-    lonlim1 -- [ymin, ymax] (longitude limits of the chunk or whole image)
-    latlim1 -- [ymin, ymax] (latitude limits of the chunk or whole image)
-    '''
-    # calculate min and max longitude and latitude
-    # lat down    lat up      lon left     lon right
-    tiletextExtremes = np.empty([len(tiletext2),6])
-    tiletextExtremes[:,0] = tiletext2[:,0]
-    tiletextExtremes[:,1] = tiletext2[:,1]
-    tiletextExtremes[:,2] = np.minimum(tiletext2[:,3], tiletext2[:,9])
-    tiletextExtremes[:,3] = np.maximum(tiletext2[:,5], tiletext2[:,7])
-    tiletextExtremes[:,4] = np.minimum(tiletext2[:,2], tiletext2[:,4])
-    tiletextExtremes[:,5] = np.maximum(tiletext2[:,6], tiletext2[:,8])
-
-    # Define the upper left tile
-    latlimtiles1UL = tiletextExtremes[np.logical_and(tiletextExtremes[:,2] <= latlim1[1], tiletextExtremes[:,3] >= latlim1[1])]#tiletext2[:,3]>=latlim[0],tiletext2[:,4]>=latlim[0],tiletext2[:,5]>=latlim[0],tiletext2[:,6]>=latlim[0],tiletext2[:,7]>=latlim[0]))]
-    latlimtilesUL = latlimtiles1UL[np.logical_and(latlimtiles1UL[:,4] <= lonlim1[0], latlimtiles1UL[:,5] >= lonlim1[0])]
-
-    # Define the lower right tile
-    latlimtiles1LR = tiletextExtremes[np.logical_and(tiletextExtremes[:,2] <= latlim1[0], tiletextExtremes[:,3] >= latlim1[0])]#tiletext2[:,3]>=latlim[0],tiletext2[:,4]>=latlim[0],tiletext2[:,5]>=latlim[0],tiletext2[:,6]>=latlim[0],tiletext2[:,7]>=latlim[0]))]
-    latlimtilesLR = latlimtiles1LR[np.logical_and(latlimtiles1LR[:,4]<=lonlim1[1],latlimtiles1LR[:,5]>=lonlim1[1])]
-
-    # Define the total tile
-    TotalTiles = np.vstack([latlimtilesUL, latlimtilesLR])
-
-    # Find the minimum horizontal and vertical tile value and the maximum horizontal and vertical tile value
-    TilesVertical = [TotalTiles[:,0].min(), TotalTiles[:,0].max()]
-    TilesHorizontal = [TotalTiles[:,1].min(), TotalTiles[:,1].max()]
-    return(TilesVertical, TilesHorizontal)
-
-def Collect_data(TilesHorizontal,TilesVertical,Date,output_folder):
+def Collect_data(TilesHorizontal,TilesVertical,Date,output_folder, hdf_library):
     '''
     This function downloads all the needed MODIS tiles from https://e4ftl01.cr.usgs.gov/MOTA/MCD43A3.006/ as a hdf file.
 
@@ -226,116 +173,130 @@ def Collect_data(TilesHorizontal,TilesVertical,Date,output_folder):
             # Download the MODIS NDVI data
             url = 'https://e4ftl01.cr.usgs.gov/MOTA/MCD43A3.006/' + Date.strftime('%Y') + '.' + Date.strftime('%m') + '.' + Date.strftime('%d') + '/'
 
-            # Get files on FTP server
-            f = urllib2.urlopen(url)
+		      # Reset the begin parameters for downloading
+            downloaded = 0
+            N=0
 
-            # Sum all the files on the server
-            soup = BeautifulSoup(f, "lxml")
-            for i in soup.findAll('a', attrs = {'href': re.compile('(?i)(hdf)$')}):
+	         # Check the library given by user
+            if hdf_library is not None:
+                os.chdir(hdf_library)
+                hdf_name = glob.glob("MCD43A3.A%s%03s.h%02dv%02d.*" %(Date.strftime('%Y'), Date.strftime('%j'), Horizontal, Vertical))
 
-                # Find the file with the wanted tile number
-                Vfile=str(i)[30:32]
-                Hfile=str(i)[27:29]
-                if int(Vfile) is int(Vertical) and int(Hfile) is int(Horizontal):
+                if len(hdf_name) == 1:
+                    hdf_file = os.path.join(hdf_library, hdf_name[0])
 
-                    # Define the whole url name
-                    full_url = urlparse.urljoin(url, i['href'])
-
-		           # Reset the begin parameters for downloading
-                    downloaded = 0
-                    N=0
-
-                    # if not downloaded try to download file
-                    while downloaded == 0:
-
-                        try:# open http and download whole .hdf
-                            nameDownload = full_url
-                            file_name = os.path.join(output_folder,nameDownload.split('/')[-1])
-                            if os.path.isfile(file_name):
-                                downloaded = 1
-                            else:
-                                x = requests.get(nameDownload, allow_redirects = False)
-                                try:
-                                    y = requests.get(x.headers['location'], auth = (username, password))
-                                except:
-                                    from requests.packages.urllib3.exceptions import InsecureRequestWarning
-                                    requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
-                                    y = requests.get(x.headers['location'], auth = (username, password), verify = False)
-                                z = open(file_name, 'wb')
-                                z.write(y.content)
-                                z.close()
-                                statinfo = os.stat(file_name)
-                                # Say that download was succesfull
-                                if int(statinfo.st_size) > 10000:
-                                     downloaded = 1
-
-                        # If download was not succesfull
-                        except:
-
-                            # Try another time
-                            N = N + 1
-
-				  # Stop trying after 10 times
-                    if N == 10:
-                        print 'Data from ' + Date.strftime('%Y-%m-%d') + ' is not available'
+                    if os.path.exists(hdf_file):
                         downloaded = 1
-                    try:
-                        # Open .hdf only band with NDVI and collect all tiles to one array
-                        dataset = gdal.Open(file_name)
-                        sdsdict = dataset.GetMetadata('SUBDATASETS')
-                        sdslist = [sdsdict[k] for k in sdsdict.keys() if '_20_NAME' in k]
-                        sds = []
+                        file_name = hdf_file
 
-                        for n in sdslist:
-                            sds.append(gdal.Open(n))
-                            full_layer = [i for i in sdslist if 'Albedo_BSA_shortwave' in i]
-                            idx = sdslist.index(full_layer[0])
-                            if Horizontal == TilesHorizontal[0] and Vertical == TilesVertical[0]:
-                                geo_t = sds[idx].GetGeoTransform()
+            if not downloaded == 1:
 
-                                # get the projection value
-                                proj = sds[idx].GetProjection()
+                # Get files on FTP server
+                f = urllib2.urlopen(url)
 
-                            data = sds[idx].ReadAsArray()
-                            countYdata = (TilesVertical[1] - TilesVertical[0] + 2) - countY
-                            DataTot_Direct[int((countYdata - 1) * 2400):int(countYdata * 2400), int((countX - 1) * 2400):int(countX * 2400)]=data
+                # Sum all the files on the server
+                soup = BeautifulSoup(f, "lxml")
+                for i in soup.findAll('a', attrs = {'href': re.compile('(?i)(hdf)$')}):
 
-                        # Open .hdf only band with NDVI and collect all tiles to one array
-                        dataset = gdal.Open(file_name)
-                        sdsdict = dataset.GetMetadata('SUBDATASETS')
-                        sdslist = [sdsdict[k] for k in sdsdict.keys() if '_30_NAME' in k]
-                        sds = []
+                    # Find the file with the wanted tile number
+                    Vfile=str(i)[30:32]
+                    Hfile=str(i)[27:29]
+                    if int(Vfile) is int(Vertical) and int(Hfile) is int(Horizontal):
 
-                        for n in sdslist:
-                            sds.append(gdal.Open(n))
-                            full_layer = [i for i in sdslist if 'Albedo_WSA_shortwave' in i]
-                            idx = sdslist.index(full_layer[0])
-                            if Horizontal == TilesHorizontal[0] and Vertical == TilesVertical[0]:
-                                geo_t = sds[idx].GetGeoTransform()
+                        # Define the whole url name
+                        full_url = urlparse.urljoin(url, i['href'])
 
-                                # get the projection value
-                                proj = sds[idx].GetProjection()
+                        # if not downloaded try to download file
+                        while downloaded == 0:
 
-                            data = sds[idx].ReadAsArray()
-                            countYdata = (TilesVertical[1] - TilesVertical[0] + 2) - countY
-                            DataTot_Diffuse[int((countYdata - 1) * 2400):int(countYdata * 2400), int((countX - 1) * 2400):int(countX * 2400)]=data
-                        del data
+                            try:# open http and download whole .hdf
+                                nameDownload = full_url
+                                file_name = os.path.join(output_folder,nameDownload.split('/')[-1])
+                                if os.path.isfile(file_name):
+                                    downloaded = 1
+                                else:
+                                    x = requests.get(nameDownload, allow_redirects = False)
+                                    try:
+                                        y = requests.get(x.headers['location'], auth = (username, password))
+                                    except:
+                                        from requests.packages.urllib3.exceptions import InsecureRequestWarning
+                                        requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
+                                        y = requests.get(x.headers['location'], auth = (username, password), verify = False)
+                                    z = open(file_name, 'wb')
+                                    z.write(y.content)
+                                    z.close()
+                                    statinfo = os.stat(file_name)
+                                    # Say that download was succesfull
+                                    if int(statinfo.st_size) > 10000:
+                                         downloaded = 1
+
+                            # If download was not succesfull
+                            except:
+
+                                # Try another time
+                                N = N + 1
+
+            				         # Stop trying after 10 times
+                                if N == 10:
+                                    print 'Data from ' + Date.strftime('%Y-%m-%d') + ' is not available'
+                                    downloaded = 1
+            try:
+                # Open .hdf only band with NDVI and collect all tiles to one array
+                dataset = gdal.Open(file_name)
+                sdsdict = dataset.GetMetadata('SUBDATASETS')
+                sdslist = [sdsdict[k] for k in sdsdict.keys() if '_20_NAME' in k]
+                sds = []
+
+                for n in sdslist:
+                    sds.append(gdal.Open(n))
+                    full_layer = [i for i in sdslist if 'Albedo_BSA_shortwave' in i]
+                    idx = sdslist.index(full_layer[0])
+                    if Horizontal == TilesHorizontal[0] and Vertical == TilesVertical[0]:
+                        geo_t = sds[idx].GetGeoTransform()
+
+                        # get the projection value
+                        proj = sds[idx].GetProjection()
+
+                    data = sds[idx].ReadAsArray()
+                    countYdata = (TilesVertical[1] - TilesVertical[0] + 2) - countY
+                    DataTot_Direct[int((countYdata - 1) * 2400):int(countYdata * 2400), int((countX - 1) * 2400):int(countX * 2400)]=data
+
+                # Open .hdf only band with NDVI and collect all tiles to one array
+                dataset = gdal.Open(file_name)
+                sdsdict = dataset.GetMetadata('SUBDATASETS')
+                sdslist = [sdsdict[k] for k in sdsdict.keys() if '_30_NAME' in k]
+                sds = []
+
+                for n in sdslist:
+                    sds.append(gdal.Open(n))
+                    full_layer = [i for i in sdslist if 'Albedo_WSA_shortwave' in i]
+                    idx = sdslist.index(full_layer[0])
+                    if Horizontal == TilesHorizontal[0] and Vertical == TilesVertical[0]:
+                        geo_t = sds[idx].GetGeoTransform()
+
+                        # get the projection value
+                        proj = sds[idx].GetProjection()
+
+                    data = sds[idx].ReadAsArray()
+                    countYdata = (TilesVertical[1] - TilesVertical[0] + 2) - countY
+                    DataTot_Diffuse[int((countYdata - 1) * 2400):int(countYdata * 2400), int((countX - 1) * 2400):int(countX * 2400)]=data
+                del data
 
 
-                    # if the tile not exists or cannot be opened, create a nan array with the right projection
-                    except:
-                        if Horizontal==TilesHorizontal[0] and Vertical==TilesVertical[0]:
-                             x1 = (TilesHorizontal[0] - 19) * 2400 * Distance
-                             x4 = (TilesVertical[0] - 9) * 2400 * -1 * Distance
-                             geo = 	[x1, Distance, 0.0, x4, 0.0, -Distance]
-                             geo_t=tuple(geo)
+            # if the tile not exists or cannot be opened, create a nan array with the right projection
+            except:
+                if Horizontal==TilesHorizontal[0] and Vertical==TilesVertical[0]:
+                     x1 = (TilesHorizontal[0] - 19) * 2400 * Distance
+                     x4 = (TilesVertical[0] - 9) * 2400 * -1 * Distance
+                     geo = 	[x1, Distance, 0.0, x4, 0.0, -Distance]
+                     geo_t=tuple(geo)
 
-                        proj='PROJCS["unnamed",GEOGCS["Unknown datum based upon the custom spheroid",DATUM["Not specified (based on custom spheroid)",SPHEROID["Custom spheroid",6371007.181,0]],PRIMEM["Greenwich",0],UNIT["degree",0.0174532925199433]],PROJECTION["Sinusoidal"],PARAMETER["longitude_of_center",0],PARAMETER["false_easting",0],PARAMETER["false_northing",0],UNIT["Meter",1]]'
-                        data=np.ones((2400,2400)) * (-9999)
-                        countYdata=(TilesVertical[1] - TilesVertical[0] + 2) - countY
-                        DataTot_Direct[(countYdata - 1) * 2400:countYdata * 2400,(countX - 1) * 2400:countX * 2400] = data
-                        DataTot_Diffuse[(countYdata - 1) * 2400:countYdata * 2400,(countX - 1) * 2400:countX * 2400] = data
-                        del data
+                proj='PROJCS["unnamed",GEOGCS["Unknown datum based upon the custom spheroid",DATUM["Not specified (based on custom spheroid)",SPHEROID["Custom spheroid",6371007.181,0]],PRIMEM["Greenwich",0],UNIT["degree",0.0174532925199433]],PROJECTION["Sinusoidal"],PARAMETER["longitude_of_center",0],PARAMETER["false_easting",0],PARAMETER["false_northing",0],UNIT["Meter",1]]'
+                data=np.ones((2400,2400)) * (-9999)
+                countYdata=(TilesVertical[1] - TilesVertical[0] + 2) - countY
+                DataTot_Direct[(countYdata - 1) * 2400:countYdata * 2400,(countX - 1) * 2400:countX * 2400] = data
+                DataTot_Diffuse[(countYdata - 1) * 2400:countYdata * 2400,(countX - 1) * 2400:countX * 2400] = data
+                del data
 
     DataTot = (0.3 * DataTot_Diffuse + 0.7 * DataTot_Direct) * 0.001
     DataTot[DataTot>5.] = -9999
